Cambios en packages y añadido algo de javadoc
