+++ src/main/java/com/uniovi/assignmentController/assigner/NotificationAssigner.java
@@ -3,4 +3,5 @@ package com.uniovi.assignmentController.assigner;
 public interface NotificationAssigner {
 
 	void assign(Long notificacionId);
+	
 }
