--- src/main/java/com/uniovi/controller/APIController.java
@@ -1,17 +1,0 @@
-package com.uniovi.controller;
-
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.uniovi.UserInfo;
-
-@RestController
-public class APIController {
-
-    @RequestMapping("/user")
-    public UserInfo user() {
-        return new UserInfo("pepe",0);
-    }
-
-}
