+++ src/main/java/com/uniovi/incidencedisplayer/MainController.java
@@ -1,4 +1,4 @@
-package com.uniovi.controller;
+package com.uniovi.incidencedisplayer;
 
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import com.uniovi.services.IncidenceService;
+import com.uniovi.incidencecontroller.services.IncidenceService;
 
 @Controller
 public class MainController {
