+++ src/main/java/com/uniovi/incidencedisplayer/OperarioController.java
@@ -1,4 +1,4 @@
-package com.uniovi.controller;
+package com.uniovi.incidencedisplayer;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -8,7 +8,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.uniovi.entities.Operario;
-import com.uniovi.services.OperarioService;
+import com.uniovi.incidencecontroller.services.OperarioService;
 
 @Controller
 public class OperarioController {
