+++ src/main/java/com/uniovi/kafkamanagement/MessageListener.java
@@ -1,4 +1,4 @@
-package com.uniovi.listeners;
+package com.uniovi.kafkamanagement;
 
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
